use template flex in node
