--- ModelInterface/Model/ReadAreaFlexInterface.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace OpenOrchestra\ModelInterface\Model;
-
-/**
- * interface ReadAreaFlexInterface
- */
-interface ReadAreaFlexInterface extends AreaFlexContainerInterface, HtmlClassContainerInterface
-{
-    /**
-     * Get htmlId
-     *
-     * @return string $htmlId
-     */
-    public function getAreaId();
-}
