--- ModelInterface/TemplateFlexEvents.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace OpenOrchestra\ModelInterface;
-
-/**
- * Class TemplateFlexEvents
- */
-class TemplateFlexEvents
-{
-    const TEMPLATE_FLEX_CREATE = 'template_flex.create';
-    const TEMPLATE_FLEX_DELETE = 'template_flex.delete';
-    const TEMPLATE_FLEX_UPDATE = 'template_flex.update';
-    const TEMPLATE_FLEX_AREA_DELETE = 'template_flex.area.delete';
-    const TEMPLATE_FLEX_AREA_UPDATE = 'template_flex.area.update';
-}
